[CHK-702] fix: amount filter bug

+++ src/main/java/it/pagopa/ecommerce/payment/methods/infrastructure/rule/AmountFilterRule.java
@@ -5,6 +5,8 @@ import it.pagopa.ecommerce.payment.methods.infrastructure.PspRepository;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
+import java.math.BigInteger;
+
 @Component
 public class AmountFilterRule implements IFilterRule {
     @Override
@@ -14,6 +16,6 @@ public class AmountFilterRule implements IFilterRule {
 
     @Override
     public Flux<PspDocument> execute(PspRepository pspRepository, Integer amount, String language, String paymentTypeCode) {
-        return pspRepository.findPspMatchAmount(amount);
+        return pspRepository.findPspMatchAmount(BigInteger.valueOf(amount));
     }
 }
